requirements.txt
Django == 3.2.10
djangorestframework == 3.13.1
pytest == 8.2.2
pytest-django == 4.8.0
djangorestframework == 3.13.1